Updated for combined files
